Move enums to only package where they're used
