+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/messages/scp/ReverseIPTagSet.java
@@ -15,7 +15,6 @@
  */
 package uk.ac.manchester.spinnaker.messages.scp;
 
-import static uk.ac.manchester.spinnaker.messages.model.IPTagCommand.SET;
 import static uk.ac.manchester.spinnaker.messages.model.IPTagFieldDefinitions.COMMAND_FIELD;
 import static uk.ac.manchester.spinnaker.messages.model.IPTagFieldDefinitions.DEST_P_FIELD;
 import static uk.ac.manchester.spinnaker.messages.model.IPTagFieldDefinitions.DEST_X_FIELD;
@@ -25,6 +24,7 @@ import static uk.ac.manchester.spinnaker.messages.model.IPTagFieldDefinitions.RE
 import static uk.ac.manchester.spinnaker.messages.model.IPTagFieldDefinitions.SDP_PORT_FIELD;
 import static uk.ac.manchester.spinnaker.messages.model.IPTagFieldDefinitions.STRIP_FIELD_BIT;
 import static uk.ac.manchester.spinnaker.messages.model.IPTagFieldDefinitions.THREE_BITS_MASK;
+import static uk.ac.manchester.spinnaker.messages.scp.IPTagCommand.SET;
 import static uk.ac.manchester.spinnaker.messages.scp.SCPCommand.CMD_IPTAG;
 
 import java.nio.ByteBuffer;
